added Feed class as abstract
